Add files via upload
Dataset and why it is interesting:
   This analysis is about Food, essential for the survival of all living organisms (a very important topic to me and for everyone else). Many of us are privileged to have the luxury to choose the food we want to eat on a daily basis. But there are still people, around the world, that struggle to get a single meal a day.       

The Food and Agriculture Organization of the United Nations provides free access to food and agriculture data for over 245 countries and territories, from the year 1961 to the most recent update (depends on the dataset). One dataset from the FAO's database is the Food Balance Sheets. It presents a comprehensive picture of the pattern of a country's food supply during a specified reference period, the last time an update was loaded to the FAO database was in 2013. The food balance sheet shows for each food item the sources of supply and its utilization. This chunk of the dataset is focused on two utilisation of each food item available:
Food - refers to the total amount of the food item available as human food during the reference period.
Feed - refers to the quantity of the food item available for feeding to the livestock and poultry during the reference period.